requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
nltk==3.8.1
konlpy==0.6.0
rank-bm25==0.2.2
schedule==1.2.0
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
numba==0.58.1
numba-scipy==0.1.0
joblib==1.3.2
concurrent-futures==3.1.1
multiprocessing-logging==0.3.4
